Quiz 0.44
Results are now fully written out to a .csv file in QuizOutput in a folder with the same name as the user, with basic information and formatting in place.

GetChecked() is now functional.

Began developing the feature to ensure the user cannot resit a quiz, which is checked by seeing if there is an associated results file for each quiz. Not currently functional