Quiz 0.51
Quiz answers for Multichoice questions are now randomly shuffled.

TrueFalse questions are now functional, and output to a results file. Attempted to simplify to single method, not currently working and commented out.

Further development on letting the user sit each quiz only once. Completed quizzes are successfully removed from the front end Listbox, however these changes are not reflected on the backend array, so it is not functional currently.