Quiz 0.61
Update the MultiSelect question type to use a List of CheckBoxes are opposed to a fixed number of pre-declared objects. Changed the constructor method to support this. Questions can now have a variable number of inputs

The MultiChoice question type has been modified to also accept a variable number of inputs, as opposed to a fixed value of 4. Checking of questions is currently not functioning due to unknown issues, and questions are no longer randomised.