Quiz 1.1
The Label lblUsername now displays the user's actual name, using studentList.csv for retrieving information.

Quizzes can now only be sat once. The quiz cannot be sat if a results file for it is already present.

Added the encoding of result output files, currently not working for unknown reasons. Created method Encode for this but currently not used, all references commented out.

Added the Decoding of Quiz and Result files within the teacher software.