Quiz 1.7
Added the functionality to display the user's final score on completion of a test, at the decision of the teacher. In the Teacher Software, a quiz can now be marked to show scoring before encoding it (given a prefix of "s-" in it's .quiz file output). Additionally, quizzes and their associated resource folder can now be deleted from the teacher software. Updated some of the information displayed in quiz questions. Images are now given slight padding on the left side. Fixed a crash when image resources were not correctly accessed.